refactor. community 리팩토링(컴포넌트 분할)
