+++ src/data/community/HomeSec1Data.js
@@ -1,0 +1,36 @@
+import mainImg from "../../assets/community/community-sec1-m.jpg";
+import slide1 from "../../assets/community/community-sec1-s1.png";
+import slide2 from "../../assets/community/community-sec1-s2.png";
+import slide3 from "../../assets/community/community-sec1-s3.jpg";
+import slide4 from "../../assets/community/community-sec1-s4.png";
+import slide5 from "../../assets/community/community-sec1-s5.png";
+import slide6 from "../../assets/community/community-sec1-s6.png";
+import slide7 from "../../assets/community/community-sec1-s7.png";
+import slide8 from "../../assets/community/community-sec1-s8.png";
+import slide9 from "../../assets/community/community-sec1-s9.png";
+import slide10 from "../../assets/community/community-sec1-s10.png";
+import slide11 from "../../assets/community/community-sec1-s11.png";
+import slide12 from "../../assets/community/community-sec1-s12.png";
+import slide13 from "../../assets/community/community-sec1-s13.png";
+import slide14 from "../../assets/community/community-sec1-s14.png";
+
+export const communityImages = {
+    mainImg,
+};
+
+export const slide = [
+    slide1,
+    slide2,
+    slide3,
+    slide4,
+    slide5,
+    slide6,
+    slide7,
+    slide8,
+    slide9,
+    slide10,
+    slide11,
+    slide12,
+    slide13,
+    slide14,
+];
