+++ src/data/community/HomeSec2Data.js
@@ -1,0 +1,53 @@
+import categoryIcon1 from "../../assets/community/community-sec2-i1.png";
+import categoryIcon2 from "../../assets/community/community-sec2-i2.png";
+import categoryIcon3 from "../../assets/community/community-sec2-i3.webp";
+import categoryIcon4 from "../../assets/community/community-sec2-i4.png";
+import categoryIcon5 from "../../assets/community/community-sec2-i5.png";
+import categoryIcon6 from "../../assets/community/community-sec2-i6.png";
+import categoryIcon7 from "../../assets/community/community-sec2-i7.png";
+import categoryIcon8 from "../../assets/community/community-sec2-i8.png";
+import categoryIcon9 from "../../assets/community/community-sec2-i9.png";
+import categoryIcon10 from "../../assets/community/community-sec2-i10.png";
+
+export const categoryIcon = [
+    {
+        img: categoryIcon1,
+        label: "가을침구특가",
+    },
+    {
+        img: categoryIcon2,
+        label: "오늘의딜",
+    },
+    {
+        img: categoryIcon3,
+        label: "집들이",
+    },
+    {
+        img: categoryIcon4,
+        label: "행운출첵",
+    },
+    {
+        img: categoryIcon5,
+        label: "패키지할인",
+    },
+    {
+        img: categoryIcon6,
+        label: "1분집들이",
+    },
+    {
+        img: categoryIcon7,
+        label: "최저가보장제",
+    },
+    {
+        img: categoryIcon8,
+        label: "원하는날도착",
+    },
+    {
+        img: categoryIcon9,
+        label: "리모델링",
+    },
+    {
+        img: categoryIcon10,
+        label: "입주청소",
+    },
+];
