+++ src/data/community/HomeSec3Data.js
@@ -1,0 +1,59 @@
+import interiorSlide1 from "../../assets/community/community-sec3-i1.gif";
+import interiorSlide2 from "../../assets/community/community-sec3-i2.jpg";
+import interiorSlide3 from "../../assets/community/community-sec3-i3.jpg";
+import interiorSlide4 from "../../assets/community/community-sec3-i4.jpg";
+import interiorSlide5 from "../../assets/community/community-sec3-i5.jpg";
+import interiorSlide6 from "../../assets/community/community-sec3-i6.jpg";
+import interiorSlide7 from "../../assets/community/community-sec3-i7.jpg";
+import interiorSlide8 from "../../assets/community/community-sec3-i8.jpg";
+import interiorSlide9 from "../../assets/community/community-sec3-i9.jpg";
+import interiorSlide10 from "../../assets/community/community-sec3-i10.jpg";
+import interiorSlide11 from "../../assets/community/community-sec3-i11.jpg";
+
+export const interiorSlide = [
+    {
+        img: interiorSlide1,
+        nickname: "ohno",
+    },
+    {
+        img: interiorSlide2,
+        nickname: "오노의집",
+    },
+    {
+        img: interiorSlide3,
+        nickname: "누룽이집",
+    },
+    {
+        img: interiorSlide4,
+        nickname: "아리의정원",
+    },
+    {
+        img: interiorSlide5,
+        nickname: "오늘의집이구망",
+    },
+    {
+        img: interiorSlide6,
+        nickname: "kozyroom",
+    },
+    {
+        img: interiorSlide7,
+        nickname: "ohno12",
+    },
+    {
+        img: interiorSlide8,
+        nickname: "웰컴투_",
+    },
+    {
+        img: interiorSlide9,
+        nickname: "막내이모홈",
+    },
+    {
+        img: interiorSlide10,
+        nickname: "오노의집",
+    },
+    {
+        img: interiorSlide11,
+        nickname: "ohno999",
+    },
+
+];
