+++ src/data/community/HomeSec4Data.js
@@ -1,0 +1,23 @@
+import TodayHouse1 from "../../assets/community/community-sec4-i1.jpg";
+import TodayHouse2 from "../../assets/community/community-sec4-i2.jpg";
+import TodayHouse3 from "../../assets/community/community-sec4-i3.jpg";
+import TodayHouse4 from "../../assets/community/community-sec4-i4.jpg";
+
+export const todayHouse = [
+    {
+        img: TodayHouse1,
+        text: "24평 구축 아파트, 일 년 내내 봄처럼 따뜻한 집으로",
+    },
+    {
+        img: TodayHouse2,
+        text: "아이와 함께하는 거실서재, 가구 배치로 분위기 바꾸는 법!",
+    },
+    {
+        img: TodayHouse3,
+        text: "취향과 육아가 공존하는 15평, 나의 작은 파리",
+    },
+    {
+        img: TodayHouse4,
+        text: "천천히 여유롭게 흐르는 떼까사의 시간",
+    },
+];
