+++ src/data/community/HomeSec5Data.js
@@ -1,0 +1,89 @@
+import categoryProduct1 from "../../assets/community/community-sec5-i1.png"
+import categoryProduct2 from "../../assets/community/community-sec5-i2.png"
+import categoryProduct3 from "../../assets/community/community-sec5-i3.png"
+import categoryProduct4 from "../../assets/community/community-sec5-i4.png"
+import categoryProduct5 from "../../assets/community/community-sec5-i5.png"
+import categoryProduct6 from "../../assets/community/community-sec5-i6.png"
+import categoryProduct7 from "../../assets/community/community-sec5-i7.png"
+import categoryProduct8 from "../../assets/community/community-sec5-i8.png"
+import categoryProduct9 from "../../assets/community/community-sec5-i9.png"
+import categoryProduct10 from "../../assets/community/community-sec5-i10.png"
+import categoryProduct11 from "../../assets/community/community-sec5-i11.png"
+import categoryProduct12 from "../../assets/community/community-sec5-i12.png"
+import categoryProduct13 from "../../assets/community/community-sec5-i13.png"
+import categoryProduct14 from "../../assets/community/community-sec5-i14.png"
+import categoryProduct15 from "../../assets/community/community-sec5-i15.png"
+import categoryProduct16 from "../../assets/community/community-sec5-i16.png"
+import categoryProduct17 from "../../assets/community/community-sec5-i17.png"
+
+export const categoryProduct = [
+    {
+        img: categoryProduct1,
+        label: "가구"
+    },
+    {
+        img: categoryProduct2,
+        label: "패브릭"
+    },
+    {
+        img: categoryProduct3,
+        label: "가전·디지털"
+    },
+    {
+        img: categoryProduct4,
+        label: "주방용품"
+    },
+    
+    {
+        img: categoryProduct5,
+        label: "식품"
+    },
+    {
+        img: categoryProduct6,
+        label: "데코·식물"
+    },
+    {
+        img: categoryProduct7,
+        label: "조명"
+    },
+    {
+        img: categoryProduct8,
+        label: "수납·정리"
+    },
+    {
+        img: categoryProduct9,
+        label: "생활용품"
+    },
+    {
+        img: categoryProduct10,
+        label: "생필품"
+    },
+    {
+        img: categoryProduct11,
+        label: "유아·아동"
+    },
+    {
+        img: categoryProduct12,
+        label: "반려동물"
+    },
+    {
+        img: categoryProduct13,
+        label: "캠핑·레저"
+    },
+    {
+        img: categoryProduct14,
+        label: "공구·DIY"
+    },
+    {
+        img: categoryProduct15,
+        label: "인테리어시공"
+    },
+    {
+        img: categoryProduct16,
+        label: "렌탈·구독"
+    },
+    {
+        img: categoryProduct17,
+        label: "장보기"
+    },
+]
