+++ src/data/community/HomeSec6Data.js
@@ -1,0 +1,21 @@
+import interiorReview1 from "../../assets/community/community-interior-i1.jpg";
+import interiorReview2 from "../../assets/community/community-interior-i2.jpg";
+import interiorReview3 from "../../assets/community/community-interior-i3.jpg";
+
+export const interiorReview = [
+    {
+        img: interiorReview1,
+        label: "코드디자인인테리어",
+        text: "제가 여러 업체 실측&상담을 한 끝에 디자인 코드를 선택했습니다 모든면에서 다 좋았지만! 정해진 예산에 최대한으로 맞춰 도와주셨구 무엇보다 소통이 정말 잘됐습니다 전체공사라 의논할게 정말 많았는데 밤낮 가리지않고 소통을 잘 해주셔서 정말 만족스러운 결과물이 나왔습니다 제 지인들께도 무조건 추천 할꼬에욧",
+    },
+    {
+        img: interiorReview2,
+        label: "요즘인테리어",
+        text: "저희는 구축 아파트를 계약해서 무조건 올수리 인테리어를 생각했던 예비 신혼 부부입니다 여러 업체를 상담했지만 인테리어는 처음이라 판단이 잘 안섰는데 현실적으로 실현 가능한 인테리어를 고민해주시고 소통도 가장 잘된다고 느꼈던 권지헌 대표님과 최종 계약을 진행했습니다 결과적으로는 기대하던 것 보다도 훨씬 인테리어가 잘 나와서 너무나도 만족스럽습니다☺️",
+    },
+    {
+        img: interiorReview3,
+        label: "ABLY_design",
+        text: "구○빈 팀장님께서 소통도 원할히 잘 해주시고 젊은 감각으로  이런저런 부자재들을 트렌디한 것으로 잘 추천해주셨습니다. 첫 미팅때부터 3D 도면 준비해서 설명해주시니 이해도 잘 되고 저희 예산에 최대한 맞춰주시려고 애쓰셨습니다. 공사 후에도 A/S 요청 편히 받아주시고 최대한 불편함 없도록 신경써 주셨습니다.",
+    },
+];
