+++ src/pages/community/HomeSec4.jsx
@@ -1,25 +1,12 @@
 import React from "react";
-import { todayHouse } from "./images";
+import { todayHouse } from "../../data/community/HomeSec4Data.js";
 import scrap from "../../assets/community/scrap.svg";
+import SectionHeader from "../../components/community/SectionHeader.jsx";
 
 const HomeSec4 = () => {
     return (
         <section className="w-full px-[60px] mb-[40px]">
-            <div className="flex justify-between items-center py-5">
-                <div>
-                    <strong className="text-xl hover:text-[#aaa]">
-                        오늘의 추천 집들이 구경해보세요🧐
-                    </strong>
-                    {/* <p className="text-sm font-light">
-                        좋아하실 만한 인테리어 콘텐츠를 추천해드려요
-                    </p> */}
-                </div>
-                <div>
-                    <span className="font-bold text-[#11A5FD]">
-                        <a href="/">더보기</a>
-                    </span>
-                </div>
-            </div>
+            <SectionHeader title={"오늘의 추천 집들이 구경해보세요🧐"} />
             <div className="grid grid-cols-4 gap-5">
                 {todayHouse.map((item, index) => {
                     return (
@@ -27,7 +14,7 @@ const HomeSec4 = () => {
                             <div className="overflow-hidden rounded-md relative">
                                 <img
                                     src={item.img}
-                                    alt=""
+                                    alt={`집들이이미지${index}`}
                                     className="h-fit object-cover hover:scale-105 transition-transform duration-300"
                                 />
                                 <div className="absolute right-2 bottom-2">
