--- src/pages/community/images.js
@@ -1,247 +1,0 @@
-import mainImg from "../../assets/community/community-sec1-m.jpg";
-import slide1 from "../../assets/community/community-sec1-s1.png";
-import slide2 from "../../assets/community/community-sec1-s2.png";
-import slide3 from "../../assets/community/community-sec1-s3.jpg";
-import slide4 from "../../assets/community/community-sec1-s4.png";
-import slide5 from "../../assets/community/community-sec1-s5.png";
-import slide6 from "../../assets/community/community-sec1-s6.png";
-import slide7 from "../../assets/community/community-sec1-s7.png";
-import slide8 from "../../assets/community/community-sec1-s8.png";
-import slide9 from "../../assets/community/community-sec1-s9.png";
-import slide10 from "../../assets/community/community-sec1-s10.png";
-import slide11 from "../../assets/community/community-sec1-s11.png";
-import slide12 from "../../assets/community/community-sec1-s12.png";
-import slide13 from "../../assets/community/community-sec1-s13.png";
-import slide14 from "../../assets/community/community-sec1-s14.png";
-import categoryIcon1 from "../../assets/community/community-sec2-i1.png";
-import categoryIcon2 from "../../assets/community/community-sec2-i2.png";
-import categoryIcon3 from "../../assets/community/community-sec2-i3.webp";
-import categoryIcon4 from "../../assets/community/community-sec2-i4.png";
-import categoryIcon5 from "../../assets/community/community-sec2-i5.png";
-import categoryIcon6 from "../../assets/community/community-sec2-i6.png";
-import categoryIcon7 from "../../assets/community/community-sec2-i7.png";
-import categoryIcon8 from "../../assets/community/community-sec2-i8.png";
-import categoryIcon9 from "../../assets/community/community-sec2-i9.png";
-import categoryIcon10 from "../../assets/community/community-sec2-i10.png";
-import interiorSlide1 from "../../assets/community/community-sec3-i1.gif";
-import interiorSlide2 from "../../assets/community/community-sec3-i2.jpg";
-import interiorSlide3 from "../../assets/community/community-sec3-i3.jpg";
-import interiorSlide4 from "../../assets/community/community-sec3-i4.jpg";
-import interiorSlide5 from "../../assets/community/community-sec3-i5.jpg";
-import interiorSlide6 from "../../assets/community/community-sec3-i6.jpg";
-import interiorSlide7 from "../../assets/community/community-sec3-i7.jpg";
-import interiorSlide8 from "../../assets/community/community-sec3-i8.jpg";
-import interiorSlide9 from "../../assets/community/community-sec3-i9.jpg";
-import interiorSlide10 from "../../assets/community/community-sec3-i10.jpg";
-import interiorSlide11 from "../../assets/community/community-sec3-i11.jpg";
-import TodayHouse1 from "../../assets/community/community-sec4-i1.jpg";
-import TodayHouse2 from "../../assets/community/community-sec4-i2.jpg";
-import TodayHouse3 from "../../assets/community/community-sec4-i3.jpg";
-import TodayHouse4 from "../../assets/community/community-sec4-i4.jpg";
-import categoryProduct1 from "../../assets/community/community-sec5-i1.png"
-import categoryProduct2 from "../../assets/community/community-sec5-i2.png"
-import categoryProduct3 from "../../assets/community/community-sec5-i3.png"
-import categoryProduct4 from "../../assets/community/community-sec5-i4.png"
-import categoryProduct5 from "../../assets/community/community-sec5-i5.png"
-import categoryProduct6 from "../../assets/community/community-sec5-i6.png"
-import categoryProduct7 from "../../assets/community/community-sec5-i7.png"
-import categoryProduct8 from "../../assets/community/community-sec5-i8.png"
-import categoryProduct9 from "../../assets/community/community-sec5-i9.png"
-import categoryProduct10 from "../../assets/community/community-sec5-i10.png"
-import categoryProduct11 from "../../assets/community/community-sec5-i11.png"
-import categoryProduct12 from "../../assets/community/community-sec5-i12.png"
-import categoryProduct13 from "../../assets/community/community-sec5-i13.png"
-import categoryProduct14 from "../../assets/community/community-sec5-i14.png"
-import categoryProduct15 from "../../assets/community/community-sec5-i15.png"
-import categoryProduct16 from "../../assets/community/community-sec5-i16.png"
-import categoryProduct17 from "../../assets/community/community-sec5-i17.png"
-import interiorReview1 from "../../assets/community/community-interior-i1.jpg"
-import interiorReview2 from "../../assets/community/community-interior-i2.jpg"
-import interiorReview3 from "../../assets/community/community-interior-i3.jpg"
-
-export const communityImages = {
-    mainImg,
-};
-
-export const slide = [
-    slide1,
-    slide2,
-    slide3,
-    slide4,
-    slide5,
-    slide6,
-    slide7,
-    slide8,
-    slide9,
-    slide10,
-    slide11,
-    slide12,
-    slide13,
-    slide14,
-];
-
-export const categoryIcon = [
-    {
-        img: categoryIcon1,
-        label: "가을침구특가",
-    },
-    {
-        img: categoryIcon2,
-        label: "오늘의딜",
-    },
-    {
-        img: categoryIcon3,
-        label: "집들이",
-    },
-    {
-        img: categoryIcon4,
-        label: "행운출첵",
-    },
-    {
-        img: categoryIcon5,
-        label: "패키지할인",
-    },
-    {
-        img: categoryIcon6,
-        label: "1분집들이",
-    },
-    {
-        img: categoryIcon7,
-        label: "최저가보장제",
-    },
-    {
-        img: categoryIcon8,
-        label: "원하는날도착",
-    },
-    {
-        img: categoryIcon9,
-        label: "리모델링",
-    },
-    {
-        img: categoryIcon10,
-        label: "입주청소",
-    },
-];
-
-export const interiorSlide = [
-    interiorSlide1,
-    interiorSlide2,
-    interiorSlide3,
-    interiorSlide4,
-    interiorSlide5,
-    interiorSlide6,
-    interiorSlide7,
-    interiorSlide8,
-    interiorSlide9,
-    interiorSlide10,
-    interiorSlide11,
-];
-
-export const todayHouse = [
-    {
-        img: TodayHouse1,
-        text: "24평 구축 아파트, 일 년 내내 봄처럼 따뜻한 집으로",
-    },
-    {
-        img: TodayHouse2,
-        text: "아이와 함께하는 거실서재, 가구 배치로 분위기 바꾸는 법!",
-    },
-    {
-        img: TodayHouse3,
-        text: "취향과 육아가 공존하는 15평, 나의 작은 파리",
-    },
-    {
-        img: TodayHouse4,
-        text: "천천히 여유롭게 흐르는 떼까사의 시간",
-    },
-];
-
-export const categoryProduct = [
-    {
-        img: categoryProduct1,
-        label: "가구"
-    },
-    {
-        img: categoryProduct2,
-        label: "패브릭"
-    },
-    {
-        img: categoryProduct3,
-        label: "가전·디지털"
-    },
-    {
-        img: categoryProduct4,
-        label: "주방용품"
-    },
-    
-    {
-        img: categoryProduct5,
-        label: "식품"
-    },
-    {
-        img: categoryProduct6,
-        label: "데코·식물"
-    },
-    {
-        img: categoryProduct7,
-        label: "조명"
-    },
-    {
-        img: categoryProduct8,
-        label: "수납·정리"
-    },
-    {
-        img: categoryProduct9,
-        label: "생활용품"
-    },
-    {
-        img: categoryProduct10,
-        label: "생필품"
-    },
-    {
-        img: categoryProduct11,
-        label: "유아·아동"
-    },
-    {
-        img: categoryProduct12,
-        label: "반려동물"
-    },
-    {
-        img: categoryProduct13,
-        label: "캠핑·레저"
-    },
-    {
-        img: categoryProduct14,
-        label: "공구·DIY"
-    },
-    {
-        img: categoryProduct15,
-        label: "인테리어시공"
-    },
-    {
-        img: categoryProduct16,
-        label: "렌탈·구독"
-    },
-    {
-        img: categoryProduct17,
-        label: "장보기"
-    },
-]
-
-export const interiorReview = [
-    {
-        img: interiorReview1,
-        label: "코드디자인인테리어",
-        text: "제가 여러 업체 실측&상담을 한 끝에 디자인 코드를 선택했습니다 모든면에서 다 좋았지만! 정해진 예산에 최대한으로 맞춰 도와주셨구 무엇보다 소통이 정말 잘됐습니다 전체공사라 의논할게 정말 많았는데 밤낮 가리지않고 소통을 잘 해주셔서 정말 만족스러운 결과물이 나왔습니다 제 지인들께도 무조건 추천 할꼬에욧"
-    },
-    {
-        img: interiorReview2,
-        label: "요즘인테리어",
-        text: "저희는 구축 아파트를 계약해서 무조건 올수리 인테리어를 생각했던 예비 신혼 부부입니다 여러 업체를 상담했지만 인테리어는 처음이라 판단이 잘 안섰는데 현실적으로 실현 가능한 인테리어를 고민해주시고 소통도 가장 잘된다고 느꼈던 권지헌 대표님과 최종 계약을 진행했습니다 결과적으로는 기대하던 것 보다도 훨씬 인테리어가 잘 나와서 너무나도 만족스럽습니다☺️"
-    },
-    {
-        img: interiorReview3,
-        label: "ABLY_design",
-        text: "구○빈 팀장님께서 소통도 원할히 잘 해주시고 젊은 감각으로  이런저런 부자재들을 트렌디한 것으로 잘 추천해주셨습니다. 첫 미팅때부터 3D 도면 준비해서 설명해주시니 이해도 잘 되고 저희 예산에 최대한 맞춰주시려고 애쓰셨습니다. 공사 후에도 A/S 요청 편히 받아주시고 최대한 불편함 없도록 신경써 주셨습니다."
-    },
-]
